these are still untested, but *should* work.
basically the udev rules will auto-mount any disk added and optionally decrypt it if it's a
luks-crypted device. afterwards it will run a post-install script which will further set up an
encrypted file. eventually another script is run in the mounted directory to finish setting up
the persistent storage.

this is very bad. because we're running a script as root on the system if any usb disk is added with
it. i need to change this to do that post-install stuff myself, once i figure out what that is.

also does the cleanup in another script to take care of the initially-mounted disk and any files
that have been loop-mounted.