Added a 404 error handler that redirects the client to the captive portal's frontpage (taking advantage of the IP tables port redirect in place by default). This answers Github ticket #85.
Fixed a subtle bug in the command-line agument parser, I accidentally added a
':' which changed how the parser interprets a request to enter debug mode.

Updated some comments.

This fix now needs tested on a live node to make sure that the port redirect
really does work the way it's supposed to.  Otherwise I'll have to update the
HTML redirect slightly (no big deal, really, but it needs tested nonetheless).

Signed-off-by: The Doctor <drwho@virtadpt.net>