Fixed captive portal return code checking - the error text was incorrect, the victim of indiscriminant copy-and-pasting.
Added code that tests to see if captive_portal.py started up correctly (meaning
that it wrote a PID file someplace) and displays the process ID in
network/done.html.  If it didn't start up it displays an error code and a
warning on done.html.

Added the error variable back to the display of network/done.html.

Signed-off-by: The Doctor <drwho@virtadpt.net>