Added more debugging instrumentation to NetworkConfiguration().  Also added some testing instrumentation that forces certain things generated by this object to known values for the purpose of determining whether or not something worked.  This involved NOT retyping some variables, which is how the next thing came to pass...
Found a couple of subtle bugs in how the names of network interfaces are typed,
which appears to have been causing problems recognizing that the captive portal
daemon started.  It appears to have worked as expected in debug and test mode
on Windbringer, the next step is to test on a live system.

Signed-off-by: The Doctor <drwho@virtadpt.net>