Fixed some comments.
Added `exit 0` at the end of every stanza in captive-portal.sh.  I wasn't sure
why things were going screwy when trying to associate a client with a mesh
node - traffic was hitting the right interface but instrumentation rules I
hacked into another copy of that script were showing that both the client's
requests and the server's (eventual) replies were getting DROPped.  I have a
suspicion that not properly terminating stanzas in the shell script that does
the heavy lifting has something to do with it.  If not, well, it doesn't hurt
to have them, either.

Added the command line option 'list' to captive-portal.sh.  That command just
dumps the currently running IP tables ruleset.  This is more to help me with
debugging right now than anything else, but it'll no doubt come in handy later.

Signed-off-by: The Doctor <drwho@virtadpt.net>