Added debugging instrumentation to Status(), starting with Status.index().
Began reworking Status.index() to be more helpful to users.  I'm taking the
suggestions made by the group of hackers in London who built a Byzantium mesh
in a pub (c.f. http://wiki.hacdc.org/index.php/User_Feedback_on_Byzantium_0.1a)
and hammered on v0.1a.  This is basically a complete rewrite of the Status()
object.

Cleaned up the formatting of the style.css file to make it more consistent.
That's just one of my quirks and doesn't really mean anything, though.

Added code to format the new DIVs I added to /index.html because I'm changing
the whole layout.

Rearranged includes/leftbar.html to more closely represent the intended
workflow of the user.

Altered /index.html to display information that is more helpful to the user
(c.f., issue #11).

Added some more helpful comments to Status() because I'll be changing the
whole shebang up.  I may as well leave some breadcrumbs to find my way home if
I have to.

Deleted the system load and swap utilization features from Status.index().

Reformatted the code for some of the simpler readouts on Status.index() to make
them more consistent with the rest of the code in that module.

Added code to pull the configurations of all of the wireless interfaces on the
node, broken down into mesh interfaces and client interfaces.  Usage stats
are formatted and displayed on /index.html.

Signed-off-by: The Doctor <drwho@virtadpt.net>