Deleted the success.html file.  It makes more sense to generate and send the HTML inside of the captive portal daemon, plus it's easier to troubleshoot.
Messed arond with the captiveportal.conf file and backed the changes out
shortly thereafter because they weren't necessary.

Adding the iOS captive portal fix to the daemon (issue #100).  It's not quite
ready yet, I have to figure out how to make CherryPy return stuff for URIs that
end in /.html/.

Signed-off-by: The Doctor <drwho@virtadpt.net>