Added some TODOs to gateways.py because I'm in the back seat of a car with limited power.
Added some testing and debugging instrumentation to Gateways(), which also
makes it possible to un-comment the variable definitions for testing.

Added a variation of enumerate_network_interfaces() to Gateways() so that
newly added network interfaces will be gateway interface candidates.  I ported
it from NetworkConfiguration() but made enough changes to it that it's not
really possible to spin it off into a utility module.

Added some debugging instrumentation to Gateways.tcpip().

Cleaned up comments that I haven't looked at in months, and can now find
everything wrong with them.

Fixed some testing instrumentation in NetworkConfiguration().

Signed-off-by: The Doctor <drwho@virtadpt.net>