Fixed comments to make more sense.
Updated conditional in Gatways.enumerate_network_interfaces() to treat USB
network devices as wired rather than wireless.  It makes more sense that way.

Added a little more debugging instrumentation.

Pulled out a flag that isn't used for anything.

Took out a conditional that ensured that wired interfaces would never have
dhcpcd run on them.

Moved a template passed to SQL queries higher in the context because it's the
same template regardless of which conditional stanza it's used in.

Signed-off-by: The Doctor <drwho@virtadpt.net>