Removed the spurious database column from the SQL queries used to generate the service list.
Removed code that also queries the list of daemons running on the node
because users really aren't supposed to see them, they're only concerned
with the web applications running on the node.

Worked out the code to generate URIs properly.

Added a test to determine whether HTTP or HTTPS was used to contact the
web server.

Signed-off-by: The Doctor <drwho@virtadpt.net>