Cleaned up some spaces in the user's service directory.
Added a column to the daemons table of services.sqlite.  This column
differentiates between services the user is allowed to know about
(Etherpad-Lite) and services the user isn't allowed to know about (SSH).

Modified the code in services.py that generates a list of daemons running
on the node for the user to check the value of the 'showtouser' column.
Depending on the value of that column for a particular service, it will
either be displayed or it won't.

Signed-off-by: The Doctor <drwho@virtadpt.net>