Updated and added comments.
Added a three-cycle sleep period at startup before going into the main loop
so that there would likely be clients associated with the node to check up
on.  This is an attempt to fix the bug in which the idle client reaper script
immediately dies because it doesn't find any clients in the output of
IP tables.

Signed-off-by: The Doctor <drwho@virtadpt.net>