Reworked NetworkConfiguration.enumerate_network_interfaces() to take as an argument the name of a pseudofile in /proc/net/ rather than having /proc/net/dev hardcoded.
Reworked NetworkConfiguration.index() to call enumerate_network_interfaces()
twice, one to generate a list of all network interfaces on the node and a
second time to generate a list of all wireless interfaces on the node.  It
then takes the difference of the two sets and uses that as the list of
wired interfaces on the node.  This is more reliable than the older method
because there was a corner case which could mess with gateway configuration.
Thank you, Python, for making array manipulation easy in a way no other
language has been able to do.

Signed-off-by: The Doctor <drwho@virtadpt.net>