Initial commit of an etc/resolv.conf.head file.
This is because, when a node becomes a gateway ifplugd runs dhcpcd to configure that network
interface.  There is a subtle problem with this: Increasingly, the IP addresses of those DNSes
cannot be MASQ'd because they're 192.168/16's, too.  DNS resolution doesn't work.

But, dhcpcd allows for an etc/resolv.conf.head file, which goes into the top part of the
resolv.conf file that it writes when it gets a lease.  This means that whatever IP addresses
we have in the resolv.conf.head file will be propagated throughout the mesh.

Signed-off-by: The Doctor <drwho@virtadpt.net>