Added `-C resolv.conf` to the evocation of dhcpcd so it won't update the /etc/resolv.conf file when it gets a lease.
Added code to ifplugd.action to add 7 public DNSes to /etc/resolv.conf after dhcpcd gets a
lease.  This is so that mesh nodes can properly resolve hostnames to IP addresses when there
is a gateway.

There are a couple of false starts that are commented out which will be deleted after this
particular boondoggle gets ironed out.

Signed-off-by: The Doctor <drwho@virtadpt.net>