Porteus v2.0 uses NetworkManager now, so that has to be disabled.  The best way to do that is an empty .desktop file.
rc.M is different in Porteus v2.0, so the easiest thing to do is upgrade
the copy we have.  One of the changes is that the loopback network interface
is initialized in rc.M instead of rc.inet1, so the easiest thing to do is
disable rc.inet1 here.  Therefore, we don't need a custom rc.inet1 script
in the repo anymore.

Signed-off-by: The Doctor <drwho@virtadpt.net>