Reformatted some comments in the autoconfiguration daemon.
Created a .desktop file that will let the user attempt to reconfigure the
node under certain circumstances.

Created a shell script that is called by sudo which, for the moment, re-runs
the configuration daemon.

Reformatted some comments in the "Will it blend?" script.  Also added code
that copies and deletes the .desktop file from ~/Desktop/.

Signed-off-by: The Doctor <drwho@virtadpt.net>