Somehow this got un-frobbed.
The manpage for avahi-daemon.conf says this about "publish-dns-services":

publish-dns-servers= Takes a comma separated list of IP  addresses  for
    unicast  DNS  servers. You can use this to announce unicast DNS servers
    via mDNS. When used in conjunction with avahi-dnsconfd  on  the  client
    side this allows DHCP-like configuration of unicast DNS servers.

This list of IP addresses was empty, meaning that no DNSes were getting
propagated when the gateway went up.  The configuration option
publish-resolv-conf-dns-servers needs to be set to "yes" to make this work.
From the manpage again:

publish-resolv-conf-dns-servers= Takes a boolean value ("yes" or "no").
    If set to "yes" avahi-daemon will publish the unicast DNS servers spec‐
    ified  in /etc/resolv.conf in addition to those specified with publish-
    dns-servers. Send avahi-daemon a SIGHUP to have it  reload  this  file.
    Defaults to "no".

When ifplugd detects that it's been connected to a hardline, it switches in
the file /etc/resolv.conf.gateway, which contains lots of public accessible
DNSes.

Please, for the love of Cain, don't switch this back, it's what causes
gateway functionality to break.

Signed-off-by: The Doctor <drwho@virtadpt.net>